# docs/btt/openapi.yaml
openapi: 3.1.0
info:
  title: Live2Text API
  description: API for Live2Text application with BTT integration
  version: 1.0.0
servers:
  - url: http://localhost:8090
    description: Local server
paths:
...
  /api/btt/select-device:
    post:
      summary: Select a device
      description: Sets the device to use for transcription
      operationId: selectDevice
      tags:
        - btt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                device:
                  type: string
                  description: The device name to select
      responses:
        '200':
          description: Device selected successfully
  /api/btt/select-language:
    post:
      summary: Select a language
      description: Sets the language to use for transcription
      operationId: selectLanguage
      tags:
        - btt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                language:
                  type: string
                  description: The language code to select
      responses:
        '200':
          description: Language selected successfully
  /api/btt/select-view-mode:
    post:
      summary: Select a view mode
      description: Sets the view mode for the application
      operationId: selectViewMode
      tags:
        - btt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                viewMode:
                  type: string
                  description: The view mode to select
      responses:
        '200':
          description: View mode selected successfully
...